Introduce BasePlugin: base class for plugins
