+++ apispec/__init__.py
@@ -2,6 +2,7 @@
 """Contains the main `APISpec` class.
 """
 from .core import APISpec, Path
+from .plugin import BasePlugin
 
 __version__ = '0.37.0'
 __author__ = 'Steven Loria, Josh Carp, and contributors'
@@ -11,4 +12,5 @@ __license__ = 'MIT'
 __all__ = [
     'APISpec',
     'Path',
+    'BasePlugin',
 ]
